Fix Quest Editor infinite loop and add MiniMap
